ADD Rendering Points && Vertex Buffer
got to the point of rendering points on screen, added a vertex buffer to
send points to vertex shader main, but functionality is still broken.
Points not exactly on the origin just dont draw for some reason...?
at this point I am convinced it is some weird byte-alignment issue.
Future me problem...