FIX stabalize gravity calculation
The opporation was mixed up so the shader was dividing by the inverse of
the distance, instead of multiplying the alread inverted distance.

The body data mapping methods were also copying position data into the
velocity buffer, which was fixed.

Gravity is still unstable on unit points, likely due to sharing
input/output positional buffers, I think duplicating the position buffer
and having and in/out buffer would stablize it to perform identical to
the CPU bound version.